Add Combinable to merge worker results firstly in master one round by one round, default setting 10 worker results will be merged.

+++ guagua-core/src/main/java/ml/shifu/guagua/util/BytableDiskList.java
@@ -207,8 +207,8 @@ public class BytableDiskList<T extends Bytable> implements AppendList<T> {
     }
 
     @Override
-    public long size() {
-        return this.count;
+    public int size() {
+        return (int)this.count;
     }
 
     /**
